+++ cpskin/core/browser/layout.py
@@ -76,9 +76,10 @@ class LayoutPolicy(base.LayoutPolicy):
 
         minisite = self.request.get('cpskin_minisite', None)
         if minisite:
+            body_class += ' in-minisite'
             if minisite.is_in_minisite_mode:
                 body_class += ' in-minisite-out-portal'
-            if minisite.is_in_portal_mode:
+            elif minisite.is_in_portal_mode:
                 body_class += ' in-minisite-in-portal'
 
         if is_main_homepage(context):